requirements.txt
audioread==3.0.1
click==8.1.3
easydict==1.10
ftfy==6.1.1
h5py==3.10.0
huggingface-hub==0.14.1
knn-cuda==0.2
laion-clap==1.1.4
librosa==0.10.1
ninja==1.11.1
numpy==1.24.0
open3d==0.17.0
opencv-python==4.7.0.72
pandas==2.0.1
pillow==9.5.0
regex==2023.5.5
soundfile==0.12.1
termcolor==2.3.0
timm==0.6.13
tokenizers==0.13.3
torch==1.13.0+cu116
torchaudio==0.13.0+cu116
torchlibrosa==0.1.0
torchvision==0.14.0+cu116
tqdm==4.65.0
transformers==4.29.1
wandb==0.15.12
wget==3.2